feat: implement Blackjack module + migrate ZCoins to long
- Models: BlackjackRound (JSON state), BlackjackTablePlayer, BlackjackMessage
- Services: BlackjackEngine (full port from PHP - bet/deal/hit/stand/double/AI dealer)
- Hub: BlackjackHub (lobby + room SignalR groups)
- Controller: BlackjackController (table CRUD, role/ready, gameplay, chat)
- Views: Blackjack lobby + BlackjackRoom (card rendering, lobby/game phases, chat)
- Models: User.ZCoins/ZCoinsFrozen and ZooCoinTransaction.Amount/Balance changed int→long

+++ WWM_ASP/Controllers/LotteryController.cs
@@ -78,8 +78,8 @@ public class LotteryController(AppDbContext db) : Controller
             if (user.AvailableZCoins < req.BetAmount)
                 return Json(new { error = "Không đủ Zoo để mua vé." });
 
-            int balBefore = user.ZCoins;
-            user.ZCoins   -= req.BetAmount;
+            long balBefore = user.ZCoins;
+            user.ZCoins    -= req.BetAmount;
             user.UpdatedAt = DateTime.UtcNow;
 
             db.ZooCoinTransactions.Add(new ZooCoinTransaction
@@ -151,8 +151,8 @@ public class LotteryController(AppDbContext db) : Controller
             if (user.AvailableZCoins < price)
                 return Json(new { error = $"Không đủ Zoo. Cần {price} Zoo." });
 
-            int balBefore = user.ZCoins;
-            user.ZCoins   -= price;
+            long balBefore = user.ZCoins;
+            user.ZCoins    -= price;
             user.UpdatedAt = DateTime.UtcNow;
 
             db.ZooCoinTransactions.Add(new ZooCoinTransaction
